fix: add images; rename team components
