+++ scriptured-prayer-components/src/components/TeamMemberSwiper.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Swiper, SwiperSlide } from "swiper/react";
 import { Navigation, Pagination, Keyboard, A11y } from "swiper/modules";
 import "swiper/css";
@@ -6,11 +5,11 @@ import "swiper/css/navigation";
 import "swiper/css/pagination";
 import "swiper/css/scrollbar";
 import "~/swiper.css";
-import Slide from "./teamCard";
-import randomGuy from "~/assets/images/random-guy.jpg";
+import follin from "~/assets/images/clark_follin.jpg";
 import randomGirl from "~/assets/images/random-girl.jpg";
+import TeamMemberCard from "./TeamMemberCard";
 
-const TeamContainer: React.FC = () => {
+const TeamMemberSwiper = () => {
   return (
     <Swiper
       className="w-full p-10 pt-5"
@@ -32,42 +31,43 @@ const TeamContainer: React.FC = () => {
       }}
     >
       <SwiperSlide>
-        <Slide
-          image={randomGuy}
-          name="The Rock"
+        <TeamMemberCard
+          image={follin}
+          name="Follin Clark"
+          position="Creator"
           description="Lorem ipsum, dolor sit amet consectetur adipisicing elit. Reprehenderit facere deleniti officiis non vero repellendus dignissimos, praesentium esse explicabo."
         />
       </SwiperSlide>
       <SwiperSlide>
-        <Slide
+        <TeamMemberCard
           image={randomGirl}
-          name="Amanda Jamanda"
+          name="Ben Clark"
           description="Lorem ipsum, dolor sit amet consectetur adipisicing elit. Reprehenderit facere deleniti officiis non vero repellendus dignissimos, praesentium esse explicabo."
         />
       </SwiperSlide>
       <SwiperSlide>
-        <Slide
-          image={randomGuy}
+        <TeamMemberCard
+          image={randomGirl}
           name="Dwayne Johnson"
           description="Lorem ipsum, dolor sit amet consectetur adipisicing elit. Reprehenderit facere deleniti officiis non vero repellendus dignissimos, praesentium esse explicabo."
         />
       </SwiperSlide>
       <SwiperSlide>
-        <Slide
+        <TeamMemberCard
           image={randomGirl}
           name="Sadison"
           description="Lorem ipsum, dolor sit amet consectetur adipisicing elit. Reprehenderit facere deleniti officiis non vero repellendus dignissimos, praesentium esse explicabo."
         />
       </SwiperSlide>
       <SwiperSlide>
-        <Slide
-          image={randomGuy}
+        <TeamMemberCard
+          image={randomGirl}
           name="The Brahma Bull"
           description="Lorem ipsum, dolor sit amet consectetur adipisicing elit. Reprehenderit facere deleniti officiis non vero repellendus dignissimos, praesentium esse explicabo."
         />
       </SwiperSlide>
       <SwiperSlide>
-        <Slide
+        <TeamMemberCard
           image={randomGirl}
           name="Evil Kneival"
           description="Lorem ipsum, dolor sit amet consectetur adipisicing elit. Reprehenderit facere deleniti officiis non vero repellendus dignissimos, praesentium esse explicabo."
@@ -78,4 +78,4 @@ const TeamContainer: React.FC = () => {
   );
 };
 
-export default TeamContainer;
+export default TeamMemberSwiper;
